delete and update comment - 73 complete

+++ graphql-prisma_v2/src/resolvers/Mutation.js
@@ -153,7 +153,15 @@ const Mutation = {
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: { id: userId },
+    });
+    if (!commentExists) {
+      throw new Error('Unable to delete comment');
+    }
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -163,7 +171,15 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: { id: userId },
+    });
+    if (!commentExists) {
+      throw new Error('Unable to update comment');
+    }
     return prisma.mutation.updateComment(
       {
         where: {
